Добавлена работа с JSON + мелкие фиксы

+++ Task_2_2_1/src/main/java/ru/nsu/stolyarov/Configuration.java
@@ -4,7 +4,7 @@ package ru.nsu.stolyarov;
  * Класс содержит в себе конфигурацию работы пиццерии.
  */
 public class Configuration {
-    public int[] bakers;
+    public long[] bakers;
     public int[] carriers;
     public int storageCapacity;
     public long workingTime;
