+++ Task_2_2_1/src/test/java/ru/nsu/stolyarov/PizzaNSUTest.java
@@ -6,7 +6,7 @@ class PizzaNSUTest {
     @Test
     void test() throws InterruptedException {
         Configuration myConf = new Configuration();
-        myConf.bakers = new int[]{300, 500, 700, 400, 250};
+        myConf.bakers = new long[]{300, 500, 700, 400, 250};
         myConf.carriers = new int[]{2, 1, 3, 1, 2};
         myConf.storageCapacity = 20;
         myConf.workingTime = 10000;
